ref: Move api types from lib/types to hooks/api/.../types.ts; ref: Refactor UsetTabScores to use SWR;
