+++ lib/actions/getScore.ts
@@ -1,5 +1,5 @@
 import { getUsetToken } from "@/lib/actions/getUserToken";
-import { Score } from "@/lib/types/Score";
+import { Score } from "@/lib/hooks/api/score/types";
 
 interface ScoreResponse {
   data: Score | null;
