+++ lib/hooks/api/beatmap/useBeatmap.ts
@@ -1,0 +1,10 @@
+"use client";
+
+import { Beatmap } from "@/lib/hooks/api/beatmap/types";
+import useSWR from "swr";
+
+export function useBeatmap(beatmapId: number) {
+  return useSWR<Beatmap>(`beatmap/${beatmapId}`, {
+    dedupingInterval: 1000 * 60 * 10,
+  });
+}
