+++ lib/hooks/api/types.ts
@@ -5,12 +5,20 @@ export enum GameMode {
   "taiko" = 1,
   "catch" = 2,
   "mania" = 3,
+
   "relaxstd" = 4,
   "relaxtaiko" = 5,
   "relaxcatch" = 6,
+
   "autopilotstd" = 8,
+
   "scorev2std" = 12,
   "scorev2taiko" = 13,
   "scorev2catch" = 14,
   "scorev2mania" = 15,
+}
+
+export interface PossibleErrorResult<T> {
+  data: any;
+  error?: string;
 }
