+++ lib/hooks/api/user/useUserGraph.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { UserGraph } from "@/lib/hooks/api/user/types";
-import { GameMode } from "@/lib/types/GameMode";
+import { GameMode } from "@/lib/hooks/api/types";
 import useSWR from "swr";
 
 export function useUserGraph(userId: number, mode: GameMode) {
