--- lib/types/BeatmapSet.ts
@@ -1,20 +1,0 @@
-import { BeatmapStatus } from "@/lib/types/BeatmapStatus";
-import { Beatmap } from "./Beatmap";
-
-export interface BeatmapSet {
-  id: number;
-  artist: string;
-  title: string;
-  creator: string;
-  creator_id: number;
-  status: BeatmapStatus;
-  last_updated: string;
-  submitted_date: string;
-  ranked_date?: string;
-  video: boolean;
-  beatmaps: Beatmap[];
-  description: string;
-  genre: string;
-  language: string;
-  tags: string[];
-}
