--- lib/types/BeatmapStatus.ts
@@ -1,9 +1,0 @@
-export enum BeatmapStatus {
-  Loved = "loved",
-  Qualified = "qualified",
-  Approved = "approved",
-  Ranked = "ranked",
-  Pending = "pending",
-  Graveyard = "graveyard",
-  WIP = "wip",
-}
