--- lib/types/PlayedBeatmap.ts
@@ -1,5 +1,0 @@
-import { Beatmap } from "./Beatmap";
-
-export interface PlayedBeatmap extends Beatmap {
-  play_count: number;
-}
