--- lib/types/PossibleErrorResult.ts
@@ -1,4 +1,0 @@
-export interface PossibleErrorResult<T> {
-  data: any;
-  error?: string;
-}
