--- lib/types/StatsSnapshot.ts
@@ -1,6 +1,0 @@
-export interface StatsSnapshot {
-  country_rank: number;
-  global_rank: number;
-  pp: number;
-  saved_at: string;
-}
