+++ lib/utils/gameModeToVanilla.ts
@@ -1,4 +1,4 @@
-import { GameMode } from "@/lib/types/GameMode";
+import { GameMode } from "@/lib/hooks/api/types";
 
 export function gameModeToVanilla(gameMode: GameMode) {
   return (gameMode % 4) as GameMode;
